+++ sdk/paymentRequest/index.js
@@ -1,0 +1,5 @@
+paymentRequest = require('./paymentRequest.js');
+
+exports.create = paymentRequest.create;
+exports.query = paymentRequest.query;
+exports.PaymentRequest = paymentRequest.PaymentRequest;